PortfolioOpt_QC/requirements.txt
# Requirements for Quantum Portfolio Optimization System

# Core D-Wave packages for quantum computing
dwave-ocean-sdk>=6.0.0
dwave-system>=1.20.0
dimod>=0.12.0

# Numerical computing and optimization
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0

# Financial data
yfinance>=0.2.0

# Utilities
python-dateutil>=2.8.0
